backend/requirements.txt
# ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
# Intelligent Policy Analyzer — Python Dependencies
# Generated from pip freeze. Pinned for reproducibility.
# ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━

# ─── Core Framework ──────────────────────────────────────────
fastapi==0.131.0
uvicorn[standard]==0.41.0
gunicorn==25.1.0
python-dotenv==1.2.1
python-multipart==0.0.22
starlette==0.52.1
h11==0.16.0
httptools==0.7.0
websockets==16.0
watchfiles==1.1.1

# ─── Database ────────────────────────────────────────────────
SQLAlchemy==2.0.46
asyncpg==0.31.0
alembic==1.18.4
greenlet==3.3.2.4

# ─── Auth + Security ────────────────────────────────────────
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0

# ─── Validation ──────────────────────────────────────────────
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
annotated-types==0.7.0

# ─── ML / NLP ────────────────────────────────────────────────
transformers==5.2.0
torch==2.10.0
safetensors==0.7.0
tokenizers==0.22.2
accelerate==1.12.0
sentence-transformers==5.2.3
scikit-learn==1.8.0
scipy==1.17.0
numpy==2.4.2
nltk==3.9.3
spacy==3.8.11
sentencepiece==0.2.1
shap==0.50.0

# ─── RAG Pipeline ────────────────────────────────────────────
langchain==1.2.10
langchain-community==0.4.1
langchain-core==1.2.17
langchain-huggingface==1.2.1
langchain-text-splitters==1.1.1
chromadb==1.0.7

# ─── Document Extraction ─────────────────────────────────────
PyMuPDF==1.27.1
python-docx==1.2.0
beautifulsoup4==4.13.4
chardet==5.2.0
Pillow==12.1.1
pytesseract==0.3.13

# ─── HTTP / Utilities ────────────────────────────────────────
httpx==0.28.1
aiofiles==25.1.0
requests==2.32.5
psutil==7.2.2
tqdm==4.67.3
joblib==1.5.3
regex==2026.2.19
PyYAML==6.0.3
Jinja2==3.1.6.1
MarkupSafe==3.0.3
